requirements.txt
gym==0.18.3
atari-py==0.2.9
numpy==1.21.2
opencv-python==4.5.3.56
pfrl==0.3.0
scikit-learn==1.0
scipy==1.7.1
seeding==0.1.0
threadpoolctl==3.0.0
torch==1.9.1
typing-extensions==3.10.0.2
matplotlib==3.4.3
dill==0.3.4
scikit-image==0.19.3
gin-config==0.5.0
dopamine-rl
pandas
pillow
pygame
captum